pyproject.toml
[project]
name = "chesshub_backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniil325",email = "potekhindaniil04@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "dishka (>=1.4.2,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "minio (>=7.2.15,<8.0.0)",
    "imgspy (>=0.2.2,<0.3.0)",
    "aiohttp (>=3.11.13,<4.0.0)",
    "boto3 (>=1.37.10,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "miniopy-async (>=1.22.0,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pyhumps (>=3.8.0,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyyaml (>=6.0.2,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.25.3"
factory-boy = "^3.3.3"

